refactor(parser): share the reparse leaf-splice guards
Move `context_admits`, `text_reads_are_inert`, `shifted_errors`, and the
grammar text-read survey out of the token tier into a new `reparse::leaf`.
They are the same question for any tier that splices a single leaf, and a
survey inside one tier's module reads as that tier's private business.

Pure move; no behavior change.
